DB/Miscs: Already applied on ArkDB 3.5.0

+++ src/server/game/Server/WorldSession.h
@@ -765,7 +765,7 @@ public:
     //Pet
     void HandlePetAction (WorldPacket & recv_data);
     void HandlePetStopAttack (WorldPacket& recv_data);
-    void HandlePetActionHelper (Unit *pet, uint64 guid1, uint16 spellid, uint16 flag, uint64 guid2);
+    void HandlePetActionHelper (Unit *pet, uint64 guid1, uint32 spellid, uint16 flag, uint64 guid2);
     void HandlePetNameQuery (WorldPacket & recv_data);
     void HandlePetSetAction (WorldPacket & recv_data);
     void HandlePetAbandon (WorldPacket & recv_data);
